azure_function_ppt/requirements.txt
azure-functions
semantic-kernel==1.30.0
openai>=1.0.0
python-dotenv>=1.0.0
python-pptx>=1.0.2